Bug 5754 Remove unused jersey dependency
The ODL project as a whole plans to eventually move to a jersey 2 based
web container solution.  This patch is an attempt to clean up jersey
related dependencies in sdni;  this patch removes an usused dependency
on jersey-client from the bgp wrapper.  The fact that the dependency
was declared and unused was discovered through performing:

> mvn dependency:analyze

Since the dependency is unused, it should be removed.

Change-Id: I152bea4a4afbbd02a8eca02d8e73e66f87a74fca
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>