Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Change-Id: Id53cfb431d1a3f22c5aa23d109a6611f5c4ac411
Issue-ID: RELENG-86
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>